backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
django-filter==23.3
Pillow 